VMS-62415: Fix defaultFps always 0
Restore per-stream defaultFps in device media capabilities and fix the secondary stream's maxFps leaking the primary/legacy camera-wide value.

+++ vms/libs/nx_vms_api/src/nx/vms/api/data/media_stream_capability.h
@@ -17,7 +17,7 @@ struct NX_VMS_API CameraStreamCapability
     float minBitrateKbps = 0;
     float maxBitrateKbps = 0;
     float defaultBitrateKbps = 0;
-    int defaultFps = 0;
+    float defaultFps = 0;
 
     /**%apidoc
      * minFps/maxFps are optional; when set, they take priority over availableFps.
